refactor(hush-cli): extract publish from pkg_cli
Move cmd_pkg_login, cmd_pkg_publish, and the private obtain_oidc_token
helper into pkg_cli/publish.rs. Both commands widened to pub(super);
obtain_oidc_token stays private. mod.rs imports trimmed: sign_package,
PkgManifest, load_or_generate_publisher_keypair, archive_file_name and
pack_source_dir_without_embedded_archives are no longer referenced
there (their consumers moved to publish.rs).
